+++ AutoSkillSettings.cs
