--- src/AutoSkill.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29009.5
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AutoSkill", "AutoSkill.csproj", "{F83C6ACF-9C5D-4B57-AEB4-83FEE47393E6}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "..\..\ExileApi\Core\Core.csproj", "{5539D732-34A7-44AB-9E28-116C3429B12A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F83C6ACF-9C5D-4B57-AEB4-83FEE47393E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F83C6ACF-9C5D-4B57-AEB4-83FEE47393E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F83C6ACF-9C5D-4B57-AEB4-83FEE47393E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F83C6ACF-9C5D-4B57-AEB4-83FEE47393E6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5539D732-34A7-44AB-9E28-116C3429B12A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5539D732-34A7-44AB-9E28-116C3429B12A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5539D732-34A7-44AB-9E28-116C3429B12A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5539D732-34A7-44AB-9E28-116C3429B12A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7BDACEB9-C10C-4F85-AD75-26DC3EE939B0}
-	EndGlobalSection
-EndGlobal
