--- src/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="SharpDX" version="4.0.1" targetFramework="net46" />
-  <package id="SharpDX.Mathematics" version="4.0.1" targetFramework="net46" />
-</packages>
